docs: add low_priority queue worker to example compose
The current example would've skipped messages on the `low_priority` queue,
which was introduced in Shopware 6.5.7.0.

My guess is, that it was intentionally left out to not break on some
older instances. As noted in the documentation:

> The transports option can only be configured with the low_priority transport
> if you are on version 6.5.7.0 or above.

https://developer.shopware.com/docs/guides/hosting/infrastructure/message-queue.html#admin-worker

It should be safe to change as that is a relatively old version (a full major
behind) and if you copy the example you probably use a current version.